chore: refactor and move functions
