+++ src/features/messages/queries/parse.ts
@@ -1,18 +1,9 @@
-import { ChainMetadata, MultiProvider } from '@hyperlane-xyz/sdk';
-
-import {
-  Message,
-  MessageStatus,
-  MessageStub,
-  WarpRouteChainAddressMap,
-  WarpRouteDetails,
-} from '../../../types';
+import { MultiProvider } from '@hyperlane-xyz/sdk';
+import { Message, MessageStatus, MessageStub } from '../../../types';
 import { logger } from '../../../utils/logger';
 import { tryUtf8DecodeBytes } from '../../../utils/string';
 import { DomainsEntry } from '../../chains/queries/fragments';
 import { isPiChain } from '../../chains/utils';
-
-import { fromWei, objKeys, parseWarpRouteMessage } from '@hyperlane-xyz/utils';
 import { postgresByteaToAddress, postgresByteaToString, postgresByteaToTxHash } from './encoding';
 import {
   MessageEntry,
@@ -42,70 +33,6 @@ export function parseMessageQueryResult(
   data: MessagesQueryResult | undefined,
 ): Message[] {
   return queryResult(multiProvider, scrapedChains, data, parseMessage);
-}
-
-function getTokenSymbolFromWarpRouteChainAddressMap(
-  chainMetadata: ChainMetadata,
-  address: Address,
-  warpRouteChainAddressMap: WarpRouteChainAddressMap,
-) {
-  const { name } = chainMetadata;
-  if (objKeys(warpRouteChainAddressMap).includes(name)) {
-    const chain = warpRouteChainAddressMap[name];
-    if (objKeys(chain).includes(address)) {
-      return chain[address];
-    }
-  }
-
-  return undefined;
-}
-
-export function parseWarpRouteDetails(
-  message: Message,
-  warpRouteChainAddressMap: WarpRouteChainAddressMap,
-): WarpRouteDetails | undefined {
-  try {
-    const {
-      body,
-      origin: { to },
-      totalPayment,
-      originMetadata,
-      destinationMetadata,
-      recipient,
-    } = message;
-
-    if (!body || !originMetadata || !destinationMetadata) return undefined;
-
-    const originTokenSymbol = getTokenSymbolFromWarpRouteChainAddressMap(
-      originMetadata,
-      to,
-      warpRouteChainAddressMap,
-    );
-    const destinationTokenSymbol = getTokenSymbolFromWarpRouteChainAddressMap(
-      destinationMetadata,
-      recipient,
-      warpRouteChainAddressMap,
-    );
-
-    // If token symbols are not found with the addresses, it means the message is not a warp route transfer
-    if (!originTokenSymbol || !destinationTokenSymbol) return undefined;
-
-    const parsedMessage = parseWarpRouteMessage(body);
-    const address = postgresByteaToAddress(parsedMessage.recipient, destinationMetadata);
-
-    return {
-      amount: fromWei(parsedMessage.amount.toString(), originMetadata.nativeToken?.decimals || 18),
-      totalPayment: fromWei(totalPayment, originMetadata.nativeToken?.decimals || 18),
-      endRecipient: address,
-      originTokenAddress: to,
-      originTokenSymbol: originTokenSymbol,
-      destinationTokenAddress: recipient,
-      destinationTokenSymbol: destinationTokenSymbol,
-    };
-  } catch (err) {
-    logger.error('Error parsing warp route details:', err);
-    return undefined;
-  }
 }
 
 function queryResult<D, M extends MessageStub>(
@@ -161,8 +88,6 @@ function parseMessageStub(
           }
         : undefined,
       isPiMsg,
-      originMetadata,
-      destinationMetadata,
     };
   } catch (error) {
     logger.error('Error parsing message stub', error);
@@ -224,8 +149,6 @@ function parseMessage(
       totalGasAmount: m.total_gas_amount.toString(),
       totalPayment: m.total_payment.toString(),
       numPayments: m.num_payments,
-      originMetadata,
-      destinationMetadata,
     };
   } catch (error) {
     logger.error('Error parsing message', error);
